Update dev environment config for web console
This was initially spotted because it was flagged as a security issue by SonarCloud

> Make sure using this hardcoded IP address is safe here.
>
> https://sonarcloud.io/project/security_hotspots?id=DEFRA_sroc-tcm-admin&hotspots=AXVNs2NhSMXK6iHBcFTo

When reviewing we spotted the current config does not match the setup we have on other services such as [waste-exemptions-back-office](https://github.com/DEFRA/waste-exemptions-back-office/blob/main/config/environments/development.rb) and [waste-carriers-back-office](https://github.com/DEFRA/waste-carriers-back-office/blob/main/config/environments/development.rb). Like the TCM we typically run our Rails apps in Vagrant boxes and the config in those projects take that into account.

So, this change updates the TCM to match those other projects. We also hope it will resolve the SonarCloud error!