Turn on translation errors
Rails views are a combination of the structure defined in the `*.html.erb` and content (also known as 'translation') from `config/locales`. It's how Rails can support multi-language and locale web sites out-of-the box.

On other Defra ruby services our convention has always been to have Rails raise an exception if a translation is missing when running in the `development` and `test` environments. We turn it off in `production` as generally the fallback is sufficient and we don't won't some wonky text to block users from being able to use the system.

We've spotted the TCM isn't set up in the same way as our other services and currently has errors turned off. This change updates the TCM config to bring it inline and see if we can't catch some missing translations!