Turn on translation errors (#471)
Rails views are a combination of the structure defined in the `*.html.erb` and content (also known as 'translation') from `config/locales`. It's how Rails can support multi-language and locale websites out-of-the-box.

On other Defra ruby services our convention has always been to have Rails raise an exception if a translation is missing when running in the `development` and `test` environments. We turn it off in `production` as generally the fallback is sufficient and we don't want some wonky text to block users from being able to use the system.

We've spotted the TCM isn't set up in the same way as our other services and currently has errors turned off. This change updates the TCM config to bring it in line and see if we can't catch some missing translations!

**Notes -  Update for Rails 6

After enabling the options saw a deprecation notice in the logs when running the tests.

> DEPRECATION WARNING: action_view.raise_on_missing_translations is deprecated and will be removed in Rails 6.2. Set i18n.raise_on_missing_translations instead.

We checked [Ruby on Rails guides](https://guides.rubyonrails.org/configuring.html#configuring-i18n) for what it says to do and updated to match

**Notes - fix missing translation

What do you know, having switched it on we have errors! So, this change also contained sorting out some missing translations.