Remove local file store
The TCM currently has functionality to replicate reading and writing things to AWS by using the local file system instead.

Our issues with this are

- if you are new to the project or haven't been using it for awhile it can be easy to forget this is the case
- it adds additional complexity and code (which is untested) for something intended to make working locally and writing tests easier

The preference of the existing tech team is

- the app behaves the same across all environments
- to stub or mock an external service rather than try and replicate it in tests
- to have less complexity and code to maintain

So, this change removes using the local file system as a stand in for AWS S3 and all the logic that supports this feature.