Remove all Charging Module API code (#469)
One of the best ways to improve a project and raise its code coverage is to delete unused code!

Back in 2020, there was a plan to update the TCM to do its charging work via the new Charging Module API. However, 2 issues arose

- the first iteration of the charging module had major performance issues and needed to be rebuilt as [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api)
- the TCM being 'tactical' meant the new tech team felt it would not be prudent to invest the massive effort of transforming it into being a UI only service especially with its current SonarCloud quality rating

Some work was done; connecting to the API and some logic to deal with bill runs (a concept the TCM doesn't have). But if we were to follow it through what has been done would need to change because **sroc-charging-module-api** has a different API. Plus code is never truly lost, we have it in our commit history.

So, this change simplifies the project by removing what is now redundant, unused code.