Add new S3 list response template
When you send a `GET` list request to an AWS S3 bucket you get an xml response. There is a command built into the AWS S3 SDK which can interpret that response which AWSFileStore depends on.

As we are aiming to stub requests to S3 instead of depending on the local file system we need a way we can mimic that response from AWS.

Using this template is our way of doing it.