Update all tests that depended on local file store
With this change all the tests that interacted with the local file system as a means of setting up a test or verifying behaviour have been updated to depend on stubbing the AWS requests and responses.

Ideally, we'd just stub/mock the services that the service under test calls. But we're aiming for as much test coverage as possible. So, by stubbing the requests, which is the furtherest away from these services under test, more of the TCM's code will be hit when the test runs.

As we improve our unit test coverage we can hopefully revisit this and move the stubbing/mocking closer to the code under test.