Move TransactionFileExporter
With the removal of the local file store we have found that some of the minitest tests are now failing. Rather than update minitest to work with webmock and stub all its S3 requests we'll take this opportunity to migrate those affected to RSpec.

Adding webmock to minitest also breaks all the integration tests making sorting out minitest a much bigger task. Migrating the tests is what we want to do and in comparison is probably of a similar effort.

So, we want to add an RSpec test. The next thing that hit us is that the class is not in fact a service. It probably should be, but currently it's just a PORO and doesn't implement the TCM service object pattern.

So, to make this clearer we move the class to the `app/lib` folder before we then starting writing tests for it.