Convert TransactionFileExporterTest to rspec
We are now having to deal with minitest tests that used to rely on the LocalFileStorage functionality in order to work.

This is the first we convert to rspec so that

- we help migrate our tests to rpsec (Defra standard) from minitest
- we keep AWS S3 request stubbing to just rspec avoiding duplication of setup

After making the change we confirmed the unit test coverage for TransactionFileExporterTest is exactly the same. But the new test covers 15% of the code compared to 11% so we at least get a bump there.