Remove jobs folder
We got rid of the need to ActiveJob based background jobs in previous commits when we removed the rependency on resque and redis.

We still have a jobs folder with an unsed base activejob class. So, this change removes the folder as part of our efforts to tidy up the TCM project.