Remove jobs folder (#473)
We got rid of the need for ActiveJob based background jobs in previous commits when we removed the dependency on Resque and Redis.

We still have a jobs folder with an unused base Activejob class. So, this change removes the folder as part of our efforts to tidy up the TCM project.