Remove local file store (#472)
The TCM currently has the functionality to replicate reading and writing things to AWS by using the local file system instead.

Our issues with this are

- if you are new to the project or haven't been using it for a while it can be easy to forget this is the case
- it adds additional complexity and code (which is untested) for something intended to make working locally and writing tests easier

The preference of the existing tech team is

- the app behaves the same across all environments
- to stub or mock an external service rather than try and replicate it in tests
- to have less complexity and code to maintain

So, this change removes using the local file system as a stand-in for AWS S3 and all the logic that supports this feature.

** Notes

- Add new S3 list response template

When you send a `GET` list request to an AWS S3 bucket you get an xml response. There is a command built into the AWS S3 SDK which can interpret that response which AWSFileStore depends on.

As we are aiming to stub requests to S3 instead of depending on the local file system we needed a way we could mimic that response from AWS. The S3 list template is our way of doing it.

- Update all tests that depended on local file store

With this change all the tests that interacted with the local file system as a means of setting up a test or verifying behaviour have been updated to depend on stubbing the AWS requests and responses.

Ideally, we'd just stub/mock the services that the service under test calls. But we're aiming for as much test coverage as possible. So, by stubbing the requests, which is the furthest away from these services under test, more of the TCM's code will be hit when the test runs.

As we improve our unit test coverage we can hopefully revisit this and move the stubbing/mocking closer to the code under test.

- Move TransactionFileExporter

With the removal of the local file store we have found that some of the minitest tests are now failing. Rather than update minitest to work with webmock and stub all its S3 requests we take this opportunity to migrate those affected to RSpec.

Adding webmock to minitest also breaks all the integration tests making sorting out minitest a much bigger task. Migrating the tests is what we want to do and in comparison is probably of a similar effort.

So, we want to add an RSpec test. The next thing that hit us is that the class is not in fact a service. It probably should be, but currently, it's just a PORO and doesn't implement the TCM service object pattern.

So, to make this clearer we moved the class to the `app/lib` folder.

- Convert TransactionFileExporterTest to rspec

After making the change we confirmed the unit test coverage for TransactionFileExporterTest is exactly the same. But the new test covers 15% of the code compared to 11% so we at least get a bump there.

- Add comment about pluck to filters

We had to research pluck() as part of migrating the tests from minitest to RSpec. Rather than forget it and have to research it again later we added this note for those not familiar with the function.