Migrate cfd_transaction_detail_presenter_test.rb
One of the things we found when we started the migration was that the existing minitest was testing methods that actually sit with the TransactionDetailPresenter. So, to make things clearer we created 2 RSpec tests and ensured they focused on only those methods on each presenter.

We also ensured all paths through the methods being tested were checked. This has led to a slight bump in coverage after the migration.