Migrate PasTransactionDetailPresenterTest
Convert the minitest PasTransactionDetailPresenterTest to rspec.

Rather than keep overriding the defaults we opted to add a trait to our `:transaction_detail` factory for PAS.

We also found all the work for one of the methods was actually done in the root TransactionDetailPresenter so we added a new test to cover that.