Fix issues with dynamic data
It's easier and clearer to assert expected behaviour when the test data is fixed. When we created our factories based on the minitest ones we also copied across some of the dynamic elements.

But as seen, our tests have immediately broken because they are expecting fixed results but the test data is dynamic.

So, this changes removes dynamic elements from the :transaction_detail in order to fix our tests without having to add additional logic to them.