Remove dead code
In an attempt to write a unit test for extract_errors() we trawled the project looking for any uses of it. Unfortunately we could find none which makes it highly likely it's dead code.

So, rather than make a guess on what on earcth `errors[]` is and what functionality it provides we delete the dead function.