Use let for variables not commands
Spotted that we really shuld be using let in a context to define what is special about that context. Instead we had been duplicating the full build transaction_detail command in each one and burying the difference in the call.

This corrects everywhere we have multiple contexts in a describe so that only the thing that differentiates the context is highlighted.