Migrate & expand Trans.DetailPresenter minitests (#474)
* Migrate Presenter minitests to RSpec

https://eaflood.atlassian.net/browse/CMEA-174

As part of our efforts to increase unit test coverage for the project, we spotted that some presenters have little to no unit test coverage. The rest have some coverage via Minitest.

This change does 2 things

- migrates the existing transaction detail tests to RSpec
- update the tests to ensure 100% coverage

Because of the number of changes and tests needed we've limited the scope of this change to just the `TransactionDetailPresenter` based tests and classes. We'll look at the others in subsequent PR's.