Add coverband support to the project
https://eaflood.atlassian.net/browse/CMEA-190

Whilst working to expand our unit test coverage we've become highly suspicious about just how much of the code is actually being used. We suspect that a number of methods and classes are actually redundant.

But without cross referencing each and every class and method name it can be hard to be sure. Plus ruby and rails magic sometimes mean that though something isn't referenced doesn't mean it isn't used.

[Coverband](https://github.com/danmayer/coverband) is

> A gem to measure production code usage, showing a counter for the number of times each line of code that is executed.

With this information we can quickly disregard code that is actually used and focus our checks on anything that registers no calls. The end result will be we can delete redundant code and thus improve our unit test coverage and maintenance overhead.