Implement coverband
Because it comes out of the box with rails support simply declaring it after the Rails gem in your Gemfile is al the config needed.

We then mount it as a rack app to allow us to see the coverage stats.

Note - we didn't follow the example to place it behind authentication. That is seen as a means to protect your source code. But as ours is open source there really is no need to.