Add coverband support to the project (#479)
https://eaflood.atlassian.net/browse/CMEA-190

Whilst working to expand our unit test coverage we've become highly suspicious about just how much of the code is actually being used. We suspect that a number of methods and classes are actually redundant.

But without cross-referencing each and every class and method name it can be hard to be sure. Plus ruby and rails magic sometimes mean that though something isn't referenced doesn't mean it isn't used.

[Coverband](https://github.com/danmayer/coverband) is

> A gem to measure production code usage, showing a counter for the number of times each line of code that is executed.

With this information, we can quickly disregard the code that is actually used and focus our checks on anything that registers no calls. The end result will be we can delete any redundant code and thus improve our unit test coverage and maintenance overhead.

** Notes 

Because it comes out of the box with rails support simply declaring it after the Rails gem in your Gemfile was all the config needed. We then mounted it as a rack app to allow us to see the coverage stats. We didn't follow the example to place it behind authentication. That is seen as a means to protect your source code. But as ours is open source there really is no need to.

We also had to add Redis to our CI workflow else we were getting errors. We initially tried to follow the same pattern as PostgreSQL by using a [service container](https://docs.github.com/en/actions/using-containerized-services/creating-redis-service-containers) but we couldn't get that to work. We're all set to use `REDIS_URL` but the service container only seems to support `REDIS_HOST` and `REDIS_PORT`. We tried a few combinations but something wasn't right. The alternate was [supercharge/redis-github-action](https://github.com/marketplace/actions/redis-server-in-github-actions) which handles installing and running Redis for you directly in the GitHub workflow environment. This did the trick. So, as we only intend this to be temporary we've left it there and gone with it.

Finally, we had to fix a test. We hadn't spotted that the new tests in the previous commit were asserting fixed values against dynamically generated data in the fixtures when we merged them.

* Fix comment