Remove /features folder
The presence of a `features/` folder indicates that at one point the previous team were using or thinking of [cucumber](https://cucumber.io/).

With no reference to the gem in the `Gemfile` or any actual features in the folder we surmise they instead opted to minitest [integration tests](https://guides.rubyonrails.org/testing.html#integration-testing).

This makes the `features/` folder redundant so we remove it with this change.