Enable view tracking in coverband
We [Add coverband support to the project](https://github.com/DEFRA/sroc-tcm-admin/pull/479) and then subsequently found out it also has support for tracking whether a view was rendered or not.

We don't care that it can't give us the usage at the line level. But it will still be incredibly informative to know which views (and therefore pages) are actually being used and more importantly, which aren't.

This change enables the feature in coverband.