Remove all Trans.DetailPresenter minitests
Reviewed all the assertions and they were just checking a value returned in a row matches what the presenter returns.

We feel our direct comparison of results in the specs does just as good a job. Plus we have 100% test coverage for the TransactionDetailsPresenters.

So, in this change we strip out any minitests which are not adding any value.