Add new traits to regime & transaction_detail
We are working on replacement specs for the last minitest presenter tests. They are just checking that the order of the rows returned matches some sorting that is done in the `TransactionGroupFilters` module.

A key thing is the call is made by `TransactionFilePresenter` which calls `regime_specific_sorter` in `TransactionGroupFilters`. This means we're going to need more control over selecting the regime which means we need to add a trait for `:cfd` rather than just having it as our default.

This change updates the factories to support this and updates the existing tests to specify the trait.