Remove broken method
The method calls `transaction_detail` on the object passed in which is expected to be a `TransactionHeader`. However, it doesn't have that property. It has `transaction_details` only.

Hence the code is broken and if we attempt to test it we get an error. A check of references and we can't find any usage of it so we delete it instead.