Update StandardJS, tidy `app/assets/javascripts`
We update StandardJS to the latest version. This version adds a rule to warn against the use of `var`, preferring `let` or `const`. However, the uglifier compressor we use cannot handle this ES6 syntax so we disable this rule in the appropriate files.

We also take the opportunity to delete `pagination.js` and `region.js` as they are entirely commented-out and therefore of no use, and remove some commented-out code from `tcm_select.js` and `view_mode.js`.