Fix dependency warnings (#487)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-161

This branch started as one to fix dependency warnings, but on investigation we found that all the production dependencies were unused!

* Remove ESLint

ESLint is causing dependency update issues so we remove it completely, with the intention of moving fully to StandardJS

* Remove `jest`

Jest is set up as the test runner, however there aren't any tests written. We therefore remove it for the time being.

* Fix linting error

It is perfectly valid to refer to the `history` object in browser JavaScript, however this causes a linting error as `history` is not defined. We therefore add a comment so Standard knows it is a global.

* Update StandardJS, tidy `app/assets/javascripts`

We update StandardJS to the latest version. This version adds a rule to warn against the use of `var`, preferring `let` or `const`. However, the uglifier compressor we use cannot handle this ES6 syntax so we disable this rule in the appropriate files.

We also take the opportunity to delete `pagination.js` and `region.js` as they are entirely commented-out and therefore of no use, and remove some commented-out code from `tcm_select.js` and `view_mode.js`.

* Remove React frontend and dependencies

We realised that the React frontend was not actually being used; we suspect this was a first pass at a frontend later replaced by the current one. We therefore remove it along with its dependencies.

* Remove webpack

Webpack was only being used to pack the React frontend so we no longer need it.

* Remove coffeescript

There aren't any coffeescript files in the TCM so we can remove this dependency.

* Remove babel files

There is a `.babelrc` file and a `babel-polyfill` dependency which seem to be left over from removing React. We therefore get rid of them.

* Remove yarn

Now that we have no production dependencies, we can be confident that yarn is no longer needed.