Add spec for TransactionHeaderPresenter (#488)
https://eaflood.atlassian.net/browse/CMEA-190

As part of our efforts to increase unit test coverage for the project, we spotted that some presenters have little to no unit test coverage.

This change adds a spec for the `TransactionHeaderPresenter` which currently has 0% coverage.

Notes - Remove broken method

The method calls `transaction_detail` on the object passed in which is expected to be a `TransactionHeader`. However, it doesn't have that property. It has `transaction_details` only.

Hence the code is broken and if we attempt to test it we get an error. A check of references and we can't find any usage of it so we delete it instead.