Remove generate endpoint
Our reading is that the endpoint is there to support being able to request the files are generated manually rather than waiting for the scheduled cron job to kick things off.

It relies on a DB value being set (we don't know why it wasn't just an env var!) which the current team have never seen set in any of our environments.

We have the means to manually run the job using rake so rather than add tests for something that is never used this change removes the redundant code and config value.