Add exclusion reasons to seeding (#502)
The [sroc-tcm-acceptance-tests](https://github.com/DEFRA/sroc-tcm-acceptance-tests/pull/48) is the temporary home of a selenium test of the TCM. Previously, it only worked if you were lucky enough to have data in the environment you were running against which matched what the test needed.

We now have some test files we can manually import before running the test. Ideally, to get a consistent run we reset the environment first (reset the DB and then re-seed). When we recently tried this though we hit a problem. When the DB is reset all exclusion reasons are deleted and the test expects some to exist.

Exclusion reasons are something managed by the user. However, they appear to have been created when the project went live and then left as is. This might explain why they were not previously included in the seeding. But we need them now to ensure our acceptance tests pass as they are critical to signing off each TCM release.

This change adds generating exclusion reasons to the seeding process. We've just gone with what we found in `production` as it appears to be a fixed list.