Add Docker support to the project
https://eaflood.atlassian.net/browse/CMEA-44

The same team currently maintains both this and the [SROC Charging Module API](https://github.com/DEFRA/sroc-charging-module-api). The Charging Module API uses Docker both locally (when working on the project) and in its AWS environment. Docker is now the standard for services at Defra and the preference for the team when working with a service.

Currently, this project is deployed direct to EC2 instances and the devteam relies on [Vagrant](https://www.vagrantup.com/) for creating an environment locally. We want to bring the TCM inline with the CHA to simplify developing and maintaining both services (we already have enough to deal with looking after 2 tech stacks!)

This change is a first pass of adding Docker support to the project based on what is necessary to get a local development environment up and running. We expect changes later once we understand more what is required to support an AWS ECS Ruby environment.