Add all our docker files
The key file here is Dockerfile. This covers everything needed to build our development and production images of the app. In conjunction with this we have `.dockerignore` which controls which files will get copied into the images during the build. Using the `.dockerignore` means we can just call `COPY . .` in the Dockerfile instead of explicitly listing everything.

Full details of what the Dockerfile is doing is included in its comments.

`docker-compose.yml` is used for local development. It sets out everything we need to get a local TCM environment up and running and acts as a working document for everything the TCM depends on.