Make puma the default web server
From rails 5 puma became the default web server (do `rails new` and you'll see it added to the Gemfile). The TCM is still using passenger in our AWS environments because this is the model our web-ops teams have followed for rails apps since we first built them.

With moving to Docker and running everything in containers we can drop that extra dependency and the additional complexity passenger brings. For reference, stats show puma is just as fast (and some argue faster) as passenger so we are not compromising performance.

So, we drop passenger from the production block and move puma into the general list.