Update PostgreSQL user env var
The env var the postgresql docker image expects to see is `POSTGRES_USER`. It makes sense to stick with what the default for postgres is so we update the name of the env var we are looking for.

Note: if we were to try and deploy this to one of our existing environments we would need to update the env var being used there as well.