Fix getting current git commit sha for Docker
The current logic assumes either that it is running in a location where the `.git` directory exists or one created during a Capistrano deployment.

That's fine for our local vagrant and AWS environments but no good for a docker based one. The `.git` will not exist in the image or be mounted at runtime and we certainly won't be using Capistrano.

So, instead we use the same solution we introduced for the SROC Charging Module API. We assume that the commit sha has been determined prior to build and added via a build arg.