Fix error for rails web console
It seems no matter what environment you run in, you always see an error from the rails web console!

This is there to be used in development to allow you to access the rails console from within the browser, normally when an error has just happened.

But for security you can only access it if the current address is in its allowed list. We were handling local and the default vagrant assigns. But now we need to handle when running in a Docker container.

That can change so there is nothing we can hard code. So, instead we have used the logic that if `LOG_TO_STDOUT=1` then we are running in DOCKER (rather than adding an explicit env var just for this reason) use IP to get the address and awk to format the result into something we can use.