Remove Procfile.dev
The [Procfile](https://devcenter.heroku.com/articles/procfile) is something used when running a Rails app on Heroku. No doubt the app was originally run there whilst the original team waited for their environments to be built.

We no longer have it running there (certainly we don't have access to anything!) and we should have deleted this awhile ago. But with making Docker our default we definately don't need this file.