Add Docker support to the project (#515)
https://eaflood.atlassian.net/browse/CMEA-44

The same team currently maintains both this and the [SROC Charging Module API](https://github.com/DEFRA/sroc-charging-module-api). The Charging Module API uses Docker both locally (when working on the project) and in its AWS environment. Docker is now the standard for services at Defra and the preference for the team when working with a service.

Currently, this project is deployed direct to EC2 instances and the dev team relies on [Vagrant](https://www.vagrantup.com/) for creating an environment locally. We want to bring the TCM in line with the CHA to simplify developing and maintaining both services (we already have enough to deal with looking after 2 tech stacks!)

This change is a first pass of adding Docker support to the project based on what is necessary to get a local development environment up and running. We expect changes later once we understand more what is required to support an AWS ECS Ruby environment.

** Notes

- Make puma the default web server

From rails 5 puma became the default web server (do `rails new` and you'll see it added to the Gemfile). The TCM is still using passenger in our AWS environments because this is the model our web-ops teams have followed for rails apps since we first built them.

With moving to Docker and running everything in containers we can drop that extra dependency and the additional complexity passenger brings. For reference, stats show puma is just as fast (and some argue faster) as passenger so we are not compromising performance.

So, we drop passenger from the production block and move puma into the general list.

- Add output messages to db/seeds.rb

When running our migration and seeding scripts within the container we got output for the DB but nothing for the seeding. Addin these `puts` statements just means we can see something is going on the terminal whilst waiting for the seeding to complete.

- Update PostgreSQL user env var

The env var the postgresql docker image expects to see is `POSTGRES_USER`. It makes sense to stick with what the default for postgres is so we update the name of the env var we are looking for.

Note: if we were to try and deploy this to one of our existing environments we would need to update the env var being used there as well.

- Fix getting current git commit sha for Docker

The current logic assumes either that it is running in a location where the `.git` directory exists or one created during a Capistrano deployment.

That's fine for our local vagrant and AWS environments but no good for a docker based one. The `.git` will not exist in the image or be mounted at runtime and we certainly won't be using Capistrano.

So, instead we use the same solution we introduced for the SROC Charging Module API. We assume that the commit sha has been determined prior to build and added via a build arg.

- Remove Procfile.dev

The [Procfile](https://devcenter.heroku.com/articles/procfile) is something used when running a Rails app on Heroku. No doubt the app was originally run there whilst the original team waited for their environments to be built.

We no longer have it running there (certainly we don't have access to anything!) and we should have deleted this awhile ago. But with making Docker our default we definately don't need this file.

- Set to always serve assets in production

The note has a bit more detail. In short, typically in production your web server will handle static assets. To keep things simple when using Docker we just let Rails handle it.