Remove references to vagrant (#528)
https://eaflood.atlassian.net/browse/CMEA-284

When we first joined the team there was no quick way to get a working TCM environment up and running. There wasn’t even much documentation.

On the Ruby Services team, we ensured each service had a [vagrant](https://www.vagrantup.com/) development environment. Vagrant is essentially a way to codify the installation and setup for your app, including all the things it depends on, for example, the database service.

This means new members can get up and running in minutes and the vagrant code acts as documentation for how to create a working environment.

This was one of the first things the new team did for the TCM. But with us now moving to Docker the Vagrant environment is defunct. Instead, we’ll be using Docker and [Docker Compose](https://docs.docker.com/compose/) to do the same thing.

This change removes all references to Vagrant.