Update logging and use JSON format
https://eaflood.atlassian.net/browse/CMEA-285

Orior to this change, the TCM used the standard Rails logger. It is very verbose but difficult to actually use when investigating issues. We access it via the AWS console (using CloudWatch) for our non-production and production environments.

```
Started GET "/auth/sign_in" for 172.19.0.1 at 2022-02-23 17:25:58 +0000
 Processing by Devise::SessionsController#new as HTML
   Rendering layout layouts/application.html.erb
   Rendering devise/sessions/new.html.erb within layouts/application
   Rendered devise/shared/_links.html.erb (Duration: 2.1ms | Allocations: 545)
   Rendered devise/sessions/new.html.erb within layouts/application (Duration: 68.4ms | Allocations: 4599)
   Rendered shared/_menu.html.erb (Duration: 1.7ms | Allocations: 339)
   Rendered layouts/_header.html.erb (Duration: 3.8ms | Allocations: 573)
   Rendered shared/_notices.html.erb (Duration: 1.6ms | Allocations: 196)
   Rendered layouts/_footer.html.erb (Duration: 1.9ms | Allocations: 104)
   Rendered layout layouts/application.html.erb (Duration: 5293.3ms | Allocations: 407460)
 Completed 200 OK in 5369ms (Views: 5335.3ms | ActiveRecord: 4.5ms | Allocations: 425177)
```

The [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) on the other hand logs everything as JSON and only outputs relevant information. Because it uses a standard JSON format it means we can leverage [Metric Filters](https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/FilterAndPatternSyntax.html) to easily find the information we need.

As part of moving the TCM to a Docker based environment we want to take this opportunity to ensure the logs are usable.