Add rails_semantic_logger gem to project
We had an initial attempt at avoiding using another dependency and creating our own log formatter. It's perfectly feasible but quickly needs to take account of a lot of data points and configurations of them (error messages, general messages etc)

So, all that logic would also need to be unit tested which meant it was making a simple task very complex. So, rather than try and re-invent the wheel we looked at 2 popular ones

- [Rails Semantic Logger](https://github.com/reidmorrison/rails_semantic_logger)
- [Lograge](https://github.com/roidrage/lograge)

We selected rails_semantic_logger in the end because

- we like convention but "Lograge is opinionated, very opinionated". We wanted something we felt we had more control over
- lograge's last release was back in 2019. There have been recent changes but this gave us concern
- it's marginal, but semantic_logger has fewer dependencies itself
- semantic_logger has detailed docs that go into what each of the config options do. They also included special notes for Puma

To be honest, there was not much in it. We tried semantic_logger first and it did everything we needed with ease so we called it a day there.