Add LOG_LEVEL entry to .env file
The example given in the semantic logging docs uses `LOG_LEVEL` mainly because this is the name of the env var Heroku uses to allow users to configure the log level for their dynos.

It makes sense so we stuck with it as the means to configure the log level used when the app is running. This is intended for local development though. Going forward, the default log level in `production` will be :info and not :warn.