Update logging and use JSON format (#529)
https://eaflood.atlassian.net/browse/CMEA-285

Prior to this change, the TCM used the standard Rails logger. It is very verbose but difficult to actually use when investigating issues. We access it via the AWS console (using CloudWatch) for our non-production and production environments.

```
Started GET "/auth/sign_in" for 172.19.0.1 at 2022-02-23 17:25:58 +0000
 Processing by Devise::SessionsController#new as HTML
   Rendering layout layouts/application.html.erb
   Rendering devise/sessions/new.html.erb within layouts/application
   Rendered devise/shared/_links.html.erb (Duration: 2.1ms | Allocations: 545)
   Rendered devise/sessions/new.html.erb within layouts/application (Duration: 68.4ms | Allocations: 4599)
   Rendered shared/_menu.html.erb (Duration: 1.7ms | Allocations: 339)
   Rendered layouts/_header.html.erb (Duration: 3.8ms | Allocations: 573)
   Rendered shared/_notices.html.erb (Duration: 1.6ms | Allocations: 196)
   Rendered layouts/_footer.html.erb (Duration: 1.9ms | Allocations: 104)
   Rendered layout layouts/application.html.erb (Duration: 5293.3ms | Allocations: 407460)
 Completed 200 OK in 5369ms (Views: 5335.3ms | ActiveRecord: 4.5ms | Allocations: 425177)
```

The [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) on the other hand logs everything as JSON and only outputs relevant information. Because it uses a standard JSON format it means we can leverage [Metric Filters](https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/FilterAndPatternSyntax.html) to easily find the information we need.

As part of moving the TCM to a Docker-based environment, we want to take this opportunity to ensure the logs are usable.

** Notes

The first thing to note; the default log level in `production` will now be :info and not :warn.

We had an initial attempt at avoiding using another dependency and creating our own log formatter. It's perfectly feasible but quickly needs to take account of a lot of data points and configurations of them (error messages, general messages etc)

So, all that logic would also need to be unit tested which meant it was making a simple task very complex. So, rather than try and re-invent the wheel we looked at 2 popular ones

- [Rails Semantic Logger](https://github.com/reidmorrison/rails_semantic_logger)
- [Lograge](https://github.com/roidrage/lograge)

We selected rails_semantic_logger in the end because

- we like convention but "Lograge is opinionated, very opinionated". We wanted something we felt we had more control over
- lograge's last release was back in 2019. There have been recent changes but this gave us concern
- it's marginal, but semantic_logger has fewer dependencies itself
- semantic_logger has detailed docs that go into what each of the config options do. They also included special notes for Puma

To be honest, there was not much in it. We tried semantic_logger first and it did everything we needed with ease so we called it a day there.

Our implementation of it does break with convention. Most docs for any logger you select assume you will only be using the logger in production. They assume you will be running the standard noisy logger in :debug mode when working locally.

We've opted to align with how we work on the [sroc-charging-module-api](https://github.com/defra/sroc-charging-module-api). There the log (output from the Docker container) we see when working locally matches what we would see in `production`. We would rather have a better understanding of what we expect to see in `production` than worry too much about conventions.

This is why all the configuration for semantic_logger is in `config/application.rb` and we removed all references to logging from `config/production.rb`.

Finally, the semantic logging docs have an example of using the env var `LOG_LEVEL`. It's named because this is the name of the env var Heroku uses to allow users to configure the log level for their dynos.

It is intended for local development. Override the env var in your local .env file to access the standard :debug level output if you needed it.